--- a/python/GafferSceneTest/SetTest.py
+++ b/python/GafferSceneTest/SetTest.py
@@ -84,7 +84,7 @@ class SetTest( GafferSceneTest.SceneTestCase ) :
 		self.assertEqual( s1["out"]["setNames"].getValue(), IECore.InternedStringVectorData( [ "setOne" ] ) )
 		self.assertEqual( s1["out"].set( "setOne" ).value.paths(), [ "/one" ] )
 
-		self.assertEqual( list( set( s2["out"]["setNames"].getValue() ) ), list( set( IECore.InternedStringVectorData( [ "setOne", "setTwo" ] ) ) ) )
+		self.assertEqual( set( list( s2["out"]["setNames"].getValue() ) ), set( list( IECore.InternedStringVectorData( [ "setOne", "setTwo" ] ) ) ) )
 		self.assertEqual( s2["out"].set( "setOne" ).value.paths(), [ "/one" ] )
 		self.assertEqual( s2["out"].set( "setTwo" ).value.paths(), [ "/two" ] )
 		self.assertTrue( s2["out"].set( "setOne", _copy = False ).isSame( s1["out"].set( "setOne", _copy = False ) ) )
@@ -227,13 +227,13 @@ class SetTest( GafferSceneTest.SceneTestCase ) :
 
 		s["name"].setValue( "shinyThings dullThings" )
 
-		self.assertEqual( list( set( s["out"]["setNames"].getValue() ) ), list( set( IECore.InternedStringVectorData( [ "shinyThings", "dullThings" ] ) ) ) )
+		self.assertEqual( set( list( s["out"]["setNames"].getValue() ) ), set( list( IECore.InternedStringVectorData( [ "shinyThings", "dullThings" ] ) ) ) )
 		self.assertEqual( set( s["out"].set( "shinyThings" ).value.paths() ), set( [ "/one", "/plane" ] ) )
 		self.assertEqual( set( s["out"].set( "dullThings" ).value.paths() ), set( [ "/one", "/plane" ] ) )
 
 		s["paths"].setValue( IECore.StringVectorData( [ "/two", "/sphere" ] ) )
 
-		self.assertEqual( list( set ( s["out"]["setNames"].getValue() ) ), list( set( IECore.InternedStringVectorData( [ "shinyThings", "dullThings" ] ) ) ) )
+		self.assertEqual( set( list( s["out"]["setNames"].getValue() ) ), set( list( IECore.InternedStringVectorData( [ "shinyThings", "dullThings" ] ) ) ) )
 		self.assertEqual( set( s["out"].set( "shinyThings" ).value.paths() ), set( [ "/two", "/sphere" ] ) )
 		self.assertEqual( set( s["out"].set( "dullThings" ).value.paths() ), set( [ "/two", "/sphere" ] ) )