--- a/src/GafferScene/Set.cpp
+++ b/src/GafferScene/Set.cpp
@@ -192,14 +192,8 @@ IECore::ConstInternedStringVectorDataPtr Set::computeSetNames( const Gaffer::Con
 {
 	ConstInternedStringVectorDataPtr inNamesData = inPlug()->setNamesPlug()->getValue();
 
-	InternedString name = namePlug()->getValue();
-	if( !name.string().size() )
-	{
-		return inNamesData;
-	}
-
-	const std::vector<InternedString> &inNames = inNamesData->readable();
-	if( std::find( inNames.begin(), inNames.end(), name ) != inNames.end() )
+	const std::string &names = namePlug()->getValue();
+	if( !names.size() )
 	{
 		return inNamesData;
 	}
@@ -209,14 +203,31 @@ IECore::ConstInternedStringVectorDataPtr Set::computeSetNames( const Gaffer::Con
 		return inNamesData;
 	}
 
-	InternedStringVectorDataPtr result = inNamesData->copy();
-	result->writable().push_back( name );
-	return result;
+	vector<InternedString> tokenizedNames;
+	Gaffer::tokenize( names, ' ', tokenizedNames );
+
+	InternedStringVectorDataPtr resultData = inNamesData->copy();
+
+	// inserting the new names into the vector
+	// while making sure we don't have duplicates
+	std::vector<InternedString> &result = resultData->writable();
+	result.reserve( result.size() + tokenizedNames.size() );
+	std::copy( tokenizedNames.begin(), tokenizedNames.end(), std::back_inserter( result ) );
+	std::sort( result.begin(), result.end() );
+	std::vector<InternedString>::iterator it;
+	it = std::unique( result.begin(), result.end() );
+	result.resize( std::distance( result.begin(), it ) );
+
+	return resultData;
 }
 
 void Set::hashSet( const IECore::InternedString &setName, const Gaffer::Context *context, const ScenePlug *parent, IECore::MurmurHash &h ) const
 {
-	if( setName != namePlug()->getValue() )
+	/// \todo Since hashSet is quite performance critical, it may be worth not tokenizing at all, and look for setName directly in namePlug()->getValue()
+	vector<InternedString> tokenizedNames;
+	Gaffer::tokenize( namePlug()->getValue(), ' ', tokenizedNames );
+
+	if( std::find( tokenizedNames.begin(), tokenizedNames.end(), setName ) == tokenizedNames.end() )
 	{
 		h = inPlug()->setPlug()->hash();
 		return;
@@ -230,7 +241,10 @@ void Set::hashSet( const IECore::InternedString &setName, const Gaffer::Context 
 
 GafferScene::ConstPathMatcherDataPtr Set::computeSet( const IECore::InternedString &setName, const Gaffer::Context *context, const ScenePlug *parent ) const
 {
-	if( InternedString( namePlug()->getValue() ) != setName )
+	vector<InternedString> tokenizedNames;
+	Gaffer::tokenize( namePlug()->getValue(), ' ', tokenizedNames );
+
+	if( std::find( tokenizedNames.begin(), tokenizedNames.end(), setName ) == tokenizedNames.end() )
 	{
 		return inPlug()->setPlug()->getValue();
 	}