You can now create multiple sets from a single Set node by adding the names
separated by a space.