@johnhaddon 2016-06-01
Thanks Ivan - it's very nice to have you pitching in on some Gaffer stuff! I've made a few comments on the grounds of performance and readability, but hopefully they shouldn't take too long to address.