@ivanimanishi 2016-06-02
Addressed @johnhaddon 's comments.

Some stats on the results:
```
On a fairly complex scene but no lookDev:
before:
  Scene generation    14.196s (wall), 64.380s (CPU)
after:
  Scene generation    13.104s (wall), 52.840s (CPU)
```

```
On the same scene, and with lots of lookDev boxes added:
before:
  Scene generation    323.650s (wall), 3292.100s (CPU)
after:
  Scene generation    217.738s (wall), 2109.990s (CPU)
```