@ivanimanishi 2016-06-02
Unfortunately, I'm also seeing some performance degradation when not using the dynamic builder, since in that case we are always adding one set at a time.
```
before:
  Scene generation    150.004s (wall), 1394.830s (CPU)
after:
  Scene generation    152.280s (wall), 1450.350s (CPU)
```