@ivanimanishi 2016-06-02
With the latest 2 commits, I've got the simple case (without the dynamic builder) to be as fast as before (actually, a bit faster).
I've also optimized the hashSet() call as suggested.

```
For the non-dynamic scene with lookDev
before:
  Scene generation    150.004s (wall), 1394.830s (CPU)
after:
  Scene generation    146.029s (wall), 1386.380s (CPU)
```