@ivanimanishi 2016-06-02
Also, the hashSet changes may have had some noticeable effect in performance:
```
On the same scene, and with lots of lookDev boxes added:
before:
  Scene generation    323.650s (wall), 3292.100s (CPU)
after:
  Scene generation    217.738s (wall), 2109.990s (CPU)
after hashSet optimization:
  Scene generation    209.082s (wall), 2018.120s (CPU)
```