@ivanimanishi 2016-06-02
Ended up optimizing the code for` computeSetNames()` in the same way as `hashSet()`.